extract PipelineItem from ProcessingTimelineComponent
