+++ graylog2-web-interface/src/components/pipelines/ProcessingTimelineComponent.tsx
@@ -18,25 +18,17 @@ import React, { useEffect, useState } from 'react';
 import styled, { css } from 'styled-components';
 import Immutable from 'immutable';
 
-import { defaultCompare as naturalSort } from 'logic/DefaultCompare';
-import { LinkContainer, Link } from 'components/common/router';
-import { Button } from 'components/bootstrap';
 import { DataTable, Spinner, PaginatedList, SearchForm, QueryHelper } from 'components/common';
-import { MetricContainer, CounterRate } from 'components/metrics';
-import Routes from 'routing/Routes';
 import { useStore } from 'stores/connect';
 import type { Stream } from 'stores/streams/StreamsStore';
 import StreamsStore from 'stores/streams/StreamsStore';
-import type { PaginatedPipelines } from 'stores/pipelines/PipelinesStore';
+import type { PaginatedPipelines, PipelineType } from 'stores/pipelines/PipelinesStore';
 import { PipelinesActions } from 'stores/pipelines/PipelinesStore';
-import { DEFAULT_PAGINATION } from 'stores/PaginationTypes';
+import { DEFAULT_PAGINATION, type Pagination } from 'stores/PaginationTypes';
 import usePaginationQueryParameter from 'hooks/usePaginationQueryParameter';
 import { PipelineConnectionsStore, PipelineConnectionsActions } from 'stores/pipelines/PipelineConnectionsStore';
-import useCurrentUser from 'hooks/useCurrentUser';
-import { isPermitted } from 'util/PermissionsMixin';
-import ButtonToolbar from 'components/bootstrap/ButtonToolbar';
 
-import PipelineConnectionsList from './PipelineConnectionsList';
+import PipelineListItem from './PipelineListItem';
 
 const StyledPaginatedList = styled(PaginatedList)`
   .pagination {
@@ -55,35 +47,6 @@ const Header = styled.div`
   display: flex;
 `;
 
-const PipelineStage = styled.div<{ $idle?: boolean }>(
-  ({ $idle, theme }) => css`
-    border: 1px solid ${theme.colors.gray[$idle ? '50px' : '70px']};
-    border-radius: 4px;
-    display: inline-block;
-    margin-right: 15px;
-    padding: 20px;
-    text-align: center;
-    width: 120px;
-    background-color: ${$idle
-      ? theme.utils.colorLevel(theme.colors.global.contentBackground, 10)
-      : theme.colors.global.contentBackground};
-  `,
-);
-
-const PipelineNameTD = styled.td`
-  max-width: 300px;
-  overflow-x: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  width: 300px;
-`;
-
-const StreamListTD = styled.td`
-  max-width: 150px;
-  width: 150px;
-  word-wrap: break-word;
-`;
-
 const PipelineFilter = ({ query, onSearch }: { query: string; onSearch: (query: string) => void }) => (
   <SearchForm
     query={query}
@@ -95,9 +58,11 @@ const PipelineFilter = ({ query, onSearch }: { query: string; onSearch: (query: 
   />
 );
 
-const _formatConnectedStreams = (streams) => streams.map((s) => s.title).join(', ');
-
-const _loadPipelines = (pagination, setLoading, setPaginatedPipelines) => {
+const _loadPipelines = (
+  pagination: Pagination,
+  setLoading: (value: boolean) => void,
+  setPaginatedPipelines: (pipelines: PaginatedPipelines | undefined) => void,
+) => {
   setLoading(true);
 
   PipelinesActions.listPaginated(pagination).then((paginatedPipelines) => {
@@ -107,7 +72,6 @@ const _loadPipelines = (pagination, setLoading, setPaginatedPipelines) => {
 };
 
 const ProcessingTimelineComponent = () => {
-  const currentUser = useCurrentUser();
   const { connections } = useStore(PipelineConnectionsStore);
   const { page, pageSize: perPage, resetPage, setPagination } = usePaginationQueryParameter();
   const [query, setQuery] = useState('');
@@ -128,7 +92,7 @@ const ProcessingTimelineComponent = () => {
     return <Spinner />;
   }
 
-  const handleSearch = (newQuery) => {
+  const handleSearch = (newQuery: string) => {
     resetPage();
     setQuery(newQuery);
   };
@@ -144,7 +108,7 @@ const ProcessingTimelineComponent = () => {
     </Header>
   );
 
-  const _headerCellFormatter = (header) => {
+  const _headerCellFormatter = (header: string) => {
     let className;
 
     if (header === 'Actions') {
@@ -154,32 +118,7 @@ const ProcessingTimelineComponent = () => {
     return <th className={className}>{header}</th>;
   };
 
-  const _formatStages = (pipeline, stages) => {
-    const stageNumbers = stages.map((stage) => stage.stage);
-
-    return pipelines
-      .map(({ stages: pipelineStages }) => pipelineStages.map(({ stage }) => stage))
-      .reduce(
-        (usedStagesAcc: number[], pipelineStages: number[]) =>
-          // Concat stages in a single array removing duplicates
-          Array.from(new Set([...usedStagesAcc, ...pipelineStages])),
-        [],
-      )
-      .sort(naturalSort)
-      .map((usedStage) => {
-        if (stageNumbers.indexOf(usedStage) === -1) {
-          return (
-            <PipelineStage key={`${pipeline.id}-stage${usedStage}`} $idle>
-              Idle
-            </PipelineStage>
-          );
-        }
-
-        return <PipelineStage key={`${pipeline.id}-stage${usedStage}`}>Stage {usedStage}</PipelineStage>;
-      });
-  };
-
-  const _deletePipeline = (pipeline) => () => {
+  const _deletePipeline = (pipeline: PipelineType) => {
     // TODO: Replace with ConfirmDialog components
     // eslint-disable-next-line no-alert
     if (window.confirm(`Do you really want to delete pipeline "${pipeline.title}"? This action cannot be undone.`)) {
@@ -189,53 +128,14 @@ const ProcessingTimelineComponent = () => {
       });
     }
   };
-
-  const _pipelineFormatter = (pipeline) => {
-    const { id, title, description, stages } = pipeline;
-
-    return (
-      <tr key={id}>
-        <PipelineNameTD>
-          <Link to={Routes.SYSTEM.PIPELINES.PIPELINE(id)} title={title}>
-            {title}
-          </Link>
-          <br />
-          {description}
-          <br />
-          <MetricContainer name={`org.graylog.plugins.pipelineprocessor.ast.Pipeline.${id}.executed`}>
-            <CounterRate prefix="Throughput:" suffix="msg/s" />
-          </MetricContainer>
-        </PipelineNameTD>
-        <StreamListTD>
-          <PipelineConnectionsList
-            pipeline={pipeline}
-            connections={connections}
-            streams={streams}
-            streamsFormatter={_formatConnectedStreams}
-            noConnectionsMessage={<em>Not connected</em>}
-          />
-        </StreamListTD>
-        <td>{_formatStages(pipeline, stages)}</td>
-        <td>
-          <ButtonToolbar>
-            <LinkContainer to={Routes.SYSTEM.PIPELINES.PIPELINE(id)}>
-              <Button disabled={!isPermitted(currentUser.permissions, 'pipeline:edit')} bsSize="xsmall">
-                Edit
-              </Button>
-            </LinkContainer>
-            <Button
-              disabled={!isPermitted(currentUser.permissions, 'pipeline:delete')}
-              bsStyle="danger"
-              bsSize="xsmall"
-              onClick={_deletePipeline(pipeline)}>
-              Delete
-            </Button>
-          </ButtonToolbar>
-        </td>
-      </tr>
-    );
-  };
-
+  const dataRowFormater = (pipelineItem: PipelineType) => (
+    <PipelineListItem
+      pipeline={pipelineItem}
+      pipelines={pipelines.toJS()}
+      connections={connections}
+      streams={streams}
+      onDeletePipeline={() => _deletePipeline(pipelineItem)} />
+  );
   const headers = ['Pipeline', 'Connected to Streams', 'Processing Timeline', 'Actions'];
 
   return (
@@ -251,7 +151,7 @@ const ProcessingTimelineComponent = () => {
           filterKeys={[]}
           filterLabel="Filter Pipelines"
           noDataText="No pipelines have been found"
-          dataRowFormatter={_pipelineFormatter}
+          dataRowFormatter={dataRowFormater}
         />
       </StyledPaginatedList>
     </div>
